Added all the Perceptron stuff that Jojic outlined
Did what the above says. I also subclassed Enemy into Unit and made an
Ally subclass that does that too. I changed the shader for Enemy1 to
being based off a 32x32 png image I made with MS Paint. The thing to
note about that in the Shader it is classified as "Sprite/Default",
which makes it code as a Texture2D, which is easy to work with. Also, if
you want to make more sprites, you need to change their import settings
when you drag them into the folder to "Type: Advanced" and make sure to
check "Read/Write Enabled". That's what lets it read the pixel values.