Added a place to choose Enemy or Ally in NeuralNode
I also fixed a bug where Units weren't being destroyed upon entering the
castle. Stephen, use the AddToTrainingSet method to add units to the
tower, and LearnUnits to finalize after everything has been added. I
think it would be cool to cost some amount of money per unit used for
training. Use the method in UnitGenerator that disables the NavMesh when
you make units to be added, and probably destroy them after learning.
Also, Franz, the UI needs a way to set the cutoff for the CannonfireAOE
(and maybe display numEnemiesInRange and numAllies).